Move config writing in separate class
